feat: moved command bl on command file (pull-signal)
