+++ packages/pull-signal/src/api/generate-api.ts
@@ -1,8 +1,5 @@
-import { writeFileSync } from "fs";
 import { Command, Option } from "commander";
-import { generateOpenApi } from "@ts-rest/open-api";
-import * as yaml from "js-yaml";
-import { contract } from "../contract/contract.js";
+import { generateApi } from "../commands/generate-api.command.js";
 
 const semanticVersionRegex = /^([1-9]\d*|0)(\.(([1-9]\d*)|0)){2}$/;
 
@@ -25,48 +22,6 @@ new Command()
     }
   })
   .action(async (options) => {
-    const { version } = options;
-    const openApiDocument = generateOpenApi(
-      contract,
-      {
-        components: {
-          securitySchemes: {
-            bearerAuth: {
-              type: "http",
-              scheme: "bearer",
-              bearerFormat: "JWT",
-              description:
-                "A bearer token in the format of a JWS and conformed to the specifications included in [RFC8725](https://tools.ietf.org/html/RFC8725).",
-            },
-          },
-        },
-
-        security: [
-          {
-            bearerAuth: [],
-          },
-        ],
-        servers: [
-          {
-            url: "/pull-signal",
-            description: "Pull signal data",
-          },
-        ],
-        info: {
-          title: "Pull signal Service API",
-          version,
-          license: {
-            name: "ISC",
-            url: "https://opensource.org/license/isc-license-txt",
-          },
-        },
-      },
-      {
-        setOperationId: true,
-      }
-    );
-
-    const fileOutputDocument = `./src/api/pull-signals_${openApiDocument.info.version}_.yaml`;
-    writeFileSync(fileOutputDocument, yaml.dump(openApiDocument));
+    generateApi(options.version);
   })
   .parse(process.argv);
